+++ src/renderer/settings/settings-preload.js
@@ -5,4 +5,5 @@ contextBridge.exposeInMainWorld('electronAPI', {
   setSettings: (settings) => ipcRenderer.send('set-settings', settings),
   closeSettingsWindow: () => ipcRenderer.send('close-settings-window'),
   openExternalLink: (url) => ipcRenderer.send('open-external-link', url),
+  reportSettingsSize: (size) => ipcRenderer.send('settings-size', size),
 });